+++ Game/src/Object/Equipment/Amulet.h
@@ -6,7 +6,7 @@
 class Amulet : public Equipment
 {
 public:
-    Amulet(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+    Amulet(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
         int physicalDefense, int magicalDefense,
         std::vector<Effect*> effects = std::vector<Effect*>{},
         std::array<int, 5> requirements = std::array<int, 5>{-1, -1, -1, -1, -1});
