+++ Game/src/Object/Equipment/Chests.cpp
@@ -1,7 +1,7 @@
 #include "Chests.h"
 #include "../../Character/Character.h"
 
-Chests::Chests(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+Chests::Chests(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
                int phyDefense, int magDefense,
                std::vector<Effect*> effects,
                std::array<int, 5> requirements)
