+++ Game/src/Object/Equipment/Equipment.h
@@ -8,7 +8,7 @@ class Character;
 class Equipment : public Item
 {
 public:
-    Equipment(short id, std::string name, const std::string& description, ItemType type, int weight, RarityType rarity, int durability,
+    Equipment(short id, std::string name, const std::string& description, ItemType type, float weight, RarityType rarity, int durability,
             int phyDefense, int magDefense,
             std::vector<Effect*> effects = std::vector<Effect*>{},
             std::array<int, 5> requirement = std::array<int, 5>{-1, -1, -1, -1, -1});
