+++ Game/src/Object/Equipment/Gloves.cpp
@@ -1,7 +1,7 @@
 #include "Gloves.h"
 #include "../../Character/Character.h"
 
-Gloves::Gloves(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+Gloves::Gloves(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
                int phyDefense, int magDefense,
                std::vector<Effect*> effects,
                std::array<int, 5> requirements)
