+++ Game/src/Object/Equipment/Gloves.h
@@ -5,7 +5,7 @@
 class Gloves : public Equipment
 {
 public:
-    Gloves(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity, int phyDefense, int magDefense, 
+    Gloves(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity, int phyDefense, int magDefense, 
             std::vector<Effect*> effects,
             std::array<int, 5> requirements);
     ~Gloves();
