+++ Game/src/Object/Equipment/Helmet.cpp
@@ -1,7 +1,7 @@
 #include "Helmet.h"
 #include "../../Character/Character.h"
 
-Helmet::Helmet(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+Helmet::Helmet(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
                int phyDefense, int magDefense,
                std::vector<Effect*> effects,
                std::array<int, 5> requirements)
