+++ Game/src/Object/Equipment/Helmet.h
@@ -7,7 +7,7 @@ class Character;
 class Helmet : public Equipment
 {
 public:
-    Helmet(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity, int phyDefense, int magDefense, 
+    Helmet(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity, int phyDefense, int magDefense, 
             std::vector<Effect*> effects,
             std::array<int, 5> requirements);
     ~Helmet();
