+++ Game/src/Object/Equipment/Leggings.cpp
@@ -1,7 +1,7 @@
 #include "Leggings.h"
 #include "../../Character/Character.h"
 
-Leggings::Leggings(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+Leggings::Leggings(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
                int phyDefense, int magDefense,
                std::vector<Effect*> effects,
                std::array<int, 5> requirements)
