+++ Game/src/Object/Equipment/Leggings.h
@@ -5,7 +5,7 @@
 class Leggings : public Equipment
 {
 public:
-    Leggings(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity, int phyDefense, int magDefense, 
+    Leggings(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity, int phyDefense, int magDefense, 
             std::vector<Effect*> effects,
             std::array<int, 5> requirements);
     ~Leggings();
