+++ Game/src/Object/Equipment/Ring.cpp
@@ -1,7 +1,7 @@
 #include "Ring.h"
 #include "../../Character/Character.h"
 
-Ring::Ring(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+Ring::Ring(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
         int phyDefense, int magDefense,
         std::vector<Effect*> effects,
         std::array<int, 5> requirements)
