+++ Game/src/Object/Equipment/Ring.h
@@ -6,7 +6,7 @@
 class Ring : public Equipment
 {    
 public:
-    Ring(short id, std::string name, const std::string& description, ItemType type, int weight, int durability, RarityType rarity,
+    Ring(short id, std::string name, const std::string& description, ItemType type, float weight, int durability, RarityType rarity,
         int phyDefense, int magDefense,
         std::vector<Effect*> effects,
         std::array<int, 5> requirements);
