+++ Game/src/Object/Misc/Misc.cpp
@@ -1,6 +1,6 @@
 #include "Misc.h"
 
-Misc::Misc(short id, std::string name, std::string description, ItemType type, int weight, RarityType rarity,
+Misc::Misc(short id, std::string name, std::string description, ItemType type, float weight, RarityType rarity,
         std::vector<Effect*> effects)
 : Item(id, name, description, type, weight, rarity, effects)
 {
