+++ Game/src/Object/Misc/Misc.h
@@ -5,7 +5,7 @@
 class Misc : public Item
 {
 public:
-    Misc(short id, std::string name, std::string description, ItemType type, int weight, RarityType rarity,
+    Misc(short id, std::string name, std::string description, ItemType type, float weight, RarityType rarity,
     std::vector<Effect*> effects = std::vector<Effect*>{});
     ~Misc();
 private:
