+++ Game/src/Object/Misc/Projectile.h
@@ -6,7 +6,7 @@
 class Projectile : public Misc
 {
 public:
-    Projectile(short id, const std::string& name, const std::string& description, ItemType type, int weight, RarityType rarity,
+    Projectile(short id, const std::string& name, const std::string& description, ItemType type, float weight, RarityType rarity,
             int damage, DamageType damageeType);
     ~Projectile();
     Projectile(const Projectile& projectile);
